util/tracing: rename FormatRecordedSpans to Recording.String()
Elevate the discoverability of FormatRecordedSpans() by making it the
stringer for a Recording. Remove the inferior stringer I had previously
added.

Release note: None
