+++ pkg/internal/client/txn_test.go
@@ -56,7 +56,7 @@ func TestTxnSnowballTrace(t *testing.T) {
 	log.Event(ctx, "txn complete")
 	sp.Finish()
 	collectedSpans := tracing.GetRecording(sp)
-	dump := tracing.FormatRecordedSpans(collectedSpans)
+	dump := collectedSpans.String()
 	// dump:
 	//    0.105ms      0.000ms    event:inside txn
 	//    0.275ms      0.171ms    event:client.Txn did AutoCommit. err: <nil>
